Added YAML front matter format for room definitions.
You can now add room definitions to the `world` directory in `.txt` format. These are written in plain text with YAML front matter. Synopsis follows:

    ---
    name: aRoom
    image: "http://example.com/image.jpg"
    exits:
      north: anotherRoom
      south: yetAnotherRoom
    ---
    You are standing in a room.

All the keys in the YAML are optional. You can even omit the front matter altogether to create a simple description-only room.

If the `name` key is omitted, the name of the room is determined from the filename ("foo.txt" creates a room called "foo" and so on);

Because URLs contain punctuation characters that are considered "special" in YAML, the value of the `image` key must be surrounded with double quotes.

**This commit modifies package.json. Be sure to reinstall your npm packages before trying it out.**